+++ cache.js
@@ -1,6 +1,23 @@
 'use strict'
 
 module.exports = [
+  // if you are customizing your runtime cache rules, please note that the
+  // first item in the runtime cache configuration array MUST be "start-url"
+  {
+    // MUST be the same as "start_url" in manifest.json
+    urlPattern: '/',
+    // use NetworkFirst or NetworkOnly if you redirect un-authenticated user to login page
+    // use StaleWhileRevalidate if you want to prompt user to reload when new version available
+    handler: 'NetworkFirst',
+    options: {
+      // don't change cache name
+      cacheName: 'start-url',
+      expiration: {
+        maxEntries: 1,
+        maxAgeSeconds: 24 * 60 * 60 // 24 hours
+      }
+    }
+  },
   {
     urlPattern: /^https:\/\/fonts\.(?:googleapis|gstatic)\.com\/.*/i,
     handler: 'CacheFirst',