+++ examples/lifecycle/pages/index.js
@@ -4,70 +4,76 @@ import Head from 'next/head'
 export default () => {
   // This hook only run once in browser after the component is rendered for the first time.
   // It has same effect as the old componentDidMount lifecycle callback.
-  useEffect(()=>{
+  useEffect(() => {
     if (typeof window !== 'undefined' && 'serviceWorker' in navigator && window.workbox !== undefined) {
+      const wb = window.workbox
       // add event listeners to handle any of PWA lifecycle event
       // https://developers.google.com/web/tools/workbox/reference-docs/latest/module-workbox-window.Workbox#events
-      window.workbox.addEventListener('installed', event => {
+      wb.addEventListener('installed', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
-      window.workbox.addEventListener('controlling', event => {
+      wb.addEventListener('controlling', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
-      window.workbox.addEventListener('activated', event => {
+      wb.addEventListener('activated', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
       // A common UX pattern for progressive web apps is to show a banner when a service worker has updated and waiting to install.
-      // NOTE: set skipWaiting to false in next.config.js pwa object
+      // NOTE: MUST set skipWaiting to false in next.config.js pwa object
       // https://developers.google.com/web/tools/workbox/guides/advanced-recipes#offer_a_page_reload_for_users
-      window.workbox.addEventListener('waiting', event => {
-        if (confirm('A new version is installed, reload to use the new version immediately?')) {
-          window.workbox.addEventListener('controlling', event => {
+      const promptNewVersionAvailable = event => {
+        // `event.wasWaitingBeforeRegister` will be false if this is the first time the updated service worker is waiting.
+        // When `event.wasWaitingBeforeRegister` is true, a previously updated service worker is still waiting.
+        // You may want to customize the UI prompt accordingly.
+        if (confirm('A newer version of this web app is available, reload to update?')) {
+          wb.addEventListener('controlling', event => {
             window.location.reload()
           })
-          window.workbox.messageSW({type: 'SKIP_WAITING'})
+
+          // Send a message to the waiting service worker, instructing it to activate.
+          wb.messageSW({ type: 'SKIP_WAITING' })
         } else {
-          // User rejected, new verion will be automatically load when user open the app next time.
+          console.log(
+            'User rejected to reload the web app, keep using old verion. New verion will be automatically load when user open the app next time.'
+          )
         }
-      })
+      }
+
+      wb.addEventListener('waiting', promptNewVersionAvailable)
+      wb.addEventListener('externalwaiting', promptNewVersionAvailable)
 
       // ISSUE - this is not working as expected, why?
       // I could only make message event listenser work when I manually add this listenser into sw.js file
-      window.workbox.addEventListener('message', event => {
+      wb.addEventListener('message', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
       /*
-      window.workbox.addEventListener('redundant', event => {
+      wb.addEventListener('redundant', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
-      window.workbox.addEventListener('externalinstalled', event => {
+      wb.addEventListener('externalinstalled', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
 
-      window.workbox.addEventListener('externalactivated', event => {
-        console.log(`Event ${event.type} is triggered.`)
-        console.log(event)
-      })
-
-      window.workbox.addEventListener('externalwaiting', event => {
+      wb.addEventListener('externalactivated', event => {
         console.log(`Event ${event.type} is triggered.`)
         console.log(event)
       })
       */
 
       // never forget to call register as auto register is turned off in next.config.js
-      window.workbox.register()
+      wb.register()
     }
   }, [])
 
@@ -76,7 +82,7 @@ export default () => {
       <Head>
         <title>next-pwa example</title>
       </Head>
-      <h1>Next.js + PWA = AWESOME!</h1>
+      <h1>Next.js + PWA = AWESOME!!</h1>
     </>
   )
 }